// lib/types/hybrid-analysis.types.ts
export interface HAThreatFeedItem {
  report_id?: string;
  md5?: string;
  sha1?: string;
  sha256?: string;
  sha512?: string;
  submit_name?: string;
  url_analysis: boolean;
  size?: number;
  mime?: string;
  type?: string;
  type_short?: string[];
  environment_id?: number;
  environment_description?: string;
  verdict?: number;
  verdict_human?: string;
}